requirements.txt
# Local dependencies — sensor simulator + dashboard
azure-eventhub>=5.11.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
streamlit>=1.32.0
plotly>=5.19.0
pandas>=2.0.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
streamlit-autorefresh>=1.0.1
python-dotenv>=1.0.0
